+++ native/react/src/components/atoms/index.ts
@@ -1,3 +1,4 @@
 export { default as Button } from "./Button";
 export { default as Icon } from "./Icon";
+export { default as Link } from "./Link";
 export { default as TextField } from "./TextField";
