+++ native/react/src/components/molecules/index.ts
@@ -1,5 +1,4 @@
 export { default as ProfileImage } from "./ProfileImage";
 export { default as ProfileUserName } from "./ProfileUserName";
 export { default as SignInForm } from "./SignInForm";
-export { default as SignInLink } from "./SignInLink";
 export { default as SignInSocialButton } from "./SignInSocialButton";
